requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
boto3==1.34.0
django-storages==1.14.2
python-decouple==3.8
cryptography>=41.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gspread==6.2.1
pytz==2023.3
celery==5.4.7
reportlab==4.0.7
openpyxl==3.1.2
xlsxwriter==3.1.9
# Floor plan analysis dependencies
opencv-python==4.8.1.78
numpy==1.24.3
scikit-image==0.21.0
pytesseract==0.3.10
easyocr==1.7.0
# Google AI (Gemini) for moving checklist generation
google-generativeai>=0.8.0